Update afsocket.c
an explanation of how to solve it would be helpful. I had to do a couple of tests to figure things out.
"tcp windows size per connection=(log_iw_size)/(max-connections). And that a arbitary value for log_iw_size can be mentioned in the source definition where max-connections is mentioned Eg: 'max-connections(5000) log_iw_size(500000)' means a tcp window size per conection is 100 bytes. If it is equal to or more than 100 bytes there won't be any warnings"